+++ NEWS.md
@@ -7,6 +7,7 @@
 * Upgrade internal JSON parsing engine for speed improvements (#1830)
 * Improved ergonomics for history prefix navigation (#2771)
 * Make columns resizable in the Environment pane (#4020)
+* Add Word Count command (#4237)
 * Add "Safe Mode" for opening sessions without profile scripts or workspace restoration (#4338)
 * PowerShell Core option in terminal (Windows-only)
 * Custom terminal shell option for Windows desktop (previously only on Mac, Linux, and server)